requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
blis==0.7.11
CacheControl==0.14.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.5
cryptography==45.0.5
cymem==2.0.11
dataclasses-json==0.6.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
filetype==1.2.0
firebase-admin==6.9.0
Flask==3.0.3
frozenlist==1.7.0
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
gunicorn==22.0.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-google-genai==2.1.6
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.4.4
language_data==1.3.0
lxml==6.0.0
marisa-trie==1.2.1
MarkupSafe==3.0.2
marshmallow==3.26.1
msgpack==1.1.1
multidict==6.6.3
murmurhash==1.0.13
mypy_extensions==1.1.0
nltk==3.8.1
numpy==1.26.4
orjson==3.10.18
packaging==24.2
pandas==2.2.2
pdfminer.six==20221105
pillow==11.3.0
preshed==3.0.10
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
PyMuPDF==1.23.22
PyMuPDFb==1.23.22
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.4.2
requests==2.32.4
requests-toolbelt==1.0.0
rsa==4.9.1
setuptools==78.1.1
six==1.17.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.41
srsly==2.5.1
tenacity==9.1.2
thinc==8.2.5
tqdm==4.67.1
typer==0.9.4
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.3.4
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.20.1
zstandard==0.23.0
google-ai-generativelanguage==0.6.18
langchain-google-genai==2.1.6